Редактирование пользователем своих данных

+++ app/Services/Images.php
@@ -20,27 +20,27 @@ class Images
             $nameWebp = $name . '.webp';
         
             $imagePath = storage_path('app/public/'.$folder.'/'. $imageName);
-            $img -> resize($width, function($constraint){
+            $img -> resize($width, $height, function($constraint){
             $constraint->aspectRatio();
             })->save($imagePath);
         
             $imagePathWebp = storage_path('app/public/'.$folder.'/'.$nameWebp);
-            $img_webp -> resize($width, function($constraint){
+            $img_webp -> resize($width, $height, function($constraint){
             $constraint->aspectRatio();
             })->save($imagePathWebp);
         
             if(isset($width_prewiew) || isset($height_prewiew)){
-                $name_preview = 'prev_'.$name;
+                $name_preview = 'prev_'.$imageName;
                 $name_preview_webp = 'prev_'.$nameWebp;
-    
-                $PrevImagePath = storage_path('app/public/'.$filder .'/'.$name_preview);
-                $image -> resize($width_prewiew, $height_prewiew, function($constraint){
+                
+                $PrevImagePath = storage_path('app/public/'.$folder .'/'.$name_preview);
+                $img -> resize($width_prewiew, $height_prewiew, function($constraint){
                     $constraint->aspectRatio();
                 })->save($PrevImagePath);
                 
     
-                $PrevImagePathWebp = storage_path('app/public/'.$filder .'/'.$name_preview_webp);
-                $webp_image -> resize($width_prewiew, $height_prewiew, function($constraint){
+                $PrevImagePathWebp = storage_path('app/public/'.$folder .'/'.$name_preview_webp);
+                $img_webp -> resize($width_prewiew, $height_prewiew, function($constraint){
                     $constraint->aspectRatio();
                 })->save($PrevImagePathWebp);
     
@@ -56,7 +56,7 @@ class Images
         if( $model->img != 'default.jpg'){
             Storage::delete('public/'.$folder.'/'.$model->img);
             Storage::delete('public/'.$folder.'/'.$model->img_webp);
-            if($model->img_preview){
+            if($model->img_preview && $model->img_preview != 'default_preview.jpg'){
                 Storage::delete('public/'.$folder.'/'.$model->img_preview);
                 Storage::delete('public/'.$folder.'/'.$model->img_preview_webp);
             }
@@ -65,7 +65,8 @@ class Images
 
     public static function updateImages($request, $model, $folder, $width=null, $height=null, $width_prewiew=null, $height_prewiew=null)
     {
+        
         Images::deleteImages($model,$folder);
-        Images::createImages($request, $folder, $width=null, $height=null, $width_prewiew=null, $height_prewiew=null);
+        Images::createImages($request, $folder, $width, $height, $width_prewiew, $height_prewiew);
     }
 }
